Turn off outputIsMappableFile when building to osx-arm64

+++ cctools/ld64/src/ld/OutputFile.cpp
@@ -3805,7 +3805,7 @@ void OutputFile::writeOutputFile(ld::Internal& state)
 			// <rdar://problem/12264302> Don't use mmap on non-hfs volumes
 #ifdef __APPLE__ // ld64-port
 			struct statfs fsInfo;
-			if ( statfs(_options.outputFilePath(), &fsInfo) != -1 ) {
+            if ( (_options.architecture() != CPU_TYPE_ARM64) && (statfs(_options.outputFilePath(), &fsInfo) != -1) ) {
 				if ( (strcmp(fsInfo.f_fstypename, "hfs") == 0) || (strcmp(fsInfo.f_fstypename, "apfs") == 0) ) {
 					(void)unlink(_options.outputFilePath());
 					outputIsMappableFile = true;
@@ -3835,7 +3835,7 @@ void OutputFile::writeOutputFile(ld::Internal& state)
 			end[1] = '\0';
 #ifdef __APPLE__ // ld64-port
 			struct statfs fsInfo;
-			if ( statfs(dirPath, &fsInfo) != -1 ) {
+            if ( (_options.architecture() != CPU_TYPE_ARM64) && (statfs(dirPath, &fsInfo) != -1) ) {
 				if ( (strcmp(fsInfo.f_fstypename, "hfs") == 0) || (strcmp(fsInfo.f_fstypename, "apfs") == 0) ) {
 					outputIsMappableFile = true;
 				}
